requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.0
    # via discord-py
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
build==1.3.0
    # via pip-tools
cffi==2.0.0
    # via pynacl
click==8.3.0
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
discord-py==2.6.4
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via build
pip-tools==7.5.1
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.23
    # via cffi
pynacl==1.6.0
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via aiosignal
wheel==0.45.1
    # via pip-tools
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
